Prueba 01 - 26-11-2024

+++ control-server.js
@@ -1,34 +1,51 @@
-const { exec } = require('child_process');
-const path = require('path');
+const { exec } = require("child_process");
+const path = require("path");
 
-const serverProcess = exec('npm run start', { cwd: path.resolve(__dirname, 'server') });
+// Verificar si NODE_ENV está configurado y si es "production"
+const isProduction = process.env.NODE_ENV === "production";
 
-serverProcess.stdout.on('data', (data) => {
+// Iniciar el servidor
+const serverProcess = exec("npm run start", {
+  cwd: path.resolve(__dirname, "server"),
+});
+
+serverProcess.stdout.on("data", (data) => {
   console.log(data);
 });
 
-serverProcess.stderr.on('data', (data) => {
+serverProcess.stderr.on("data", (data) => {
   console.error(data);
 });
 
-serverProcess.on('close', (code) => {
+serverProcess.on("close", (code) => {
   console.log(`Server process exited with code ${code}`);
 });
 
-process.on('SIGINT', () => {
-  console.log('Stopping the server...');
+console.log(`Running in ${isProduction ? "production" : "development"} mode`);
 
-  // Ejecutar el comando para deshacer los seeders
-  exec('npx sequelize-cli db:seed:undo:all', { cwd: path.resolve(__dirname, 'server') }, (error, stdout, stderr) => {
-    if (error) {
-      console.error(`Error executing undo seeders: ${error}`);
-      return;
-    }
-    console.log(stdout);
-    console.error(stderr);
+// Solo manejar SIGINT y ejecutar undo seeders si no estamos en producción
+if (!isProduction) {
+  process.on("SIGINT", () => {
+    console.log("Stopping the server...");
 
-    // Terminar el proceso del servidor
-    serverProcess.kill();
-    process.exit();
+    // Ejecutar el comando para deshacer los seeders
+    exec(
+      "npx sequelize-cli db:seed:undo:all",
+      { cwd: path.resolve(__dirname, "server") },
+      (error, stdout, stderr) => {
+        if (error) {
+          console.error(`Error executing undo seeders: ${error}`);
+          return;
+        }
+        console.log(stdout);
+        console.error(stderr);
+
+        // Terminar el proceso del servidor
+        serverProcess.kill();
+        process.exit();
+      }
+    );
   });
-});
+} else {
+  console.log("Running in production mode, skipping seed undo operation.");
+}
